+++ fsbp-fix/bucket-utils/bucketblocker.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 import (
 	"context"
@@ -7,16 +7,11 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/securityhub"
-	"github.com/guardian/fsbp-tools/bucket-blocker/utils"
 	common "github.com/guardian/fsbp-tools/common"
 )
 
-func main() {
-
-	ctx := context.Background()
-	args := utils.ParseArgs()
-
-	cfg, err := common.LoadDefaultConfig(ctx, args.Profile, args.Region)
+func BucketBlocker(ctx context.Context, profile string, region string, bucketCount int, exclusions []string, execute bool) {
+	cfg, err := common.LoadDefaultConfig(ctx, profile, region)
 	if err != nil {
 		log.Fatalf("unable to load SDK config, %v", err)
 	}
@@ -24,10 +19,10 @@ func main() {
 	securityHubClient := securityhub.NewFromConfig(cfg)
 	s3Client := s3.NewFromConfig(cfg)
 	cfnClient := cloudformation.NewFromConfig(cfg)
-	bucketsToBlock, err := utils.FindBucketsToBlock(ctx, securityHubClient, s3Client, cfnClient, args.BucketCount, args.Exclusions)
+	bucketsToBlock, err := FindBucketsToBlock(ctx, securityHubClient, s3Client, cfnClient, int32(bucketCount), exclusions)
 	if err != nil {
 		log.Fatalf("Error working out which buckets need blocking: %v", err)
 	}
 
-	utils.BlockBuckets(ctx, s3Client, bucketsToBlock, args.Execute)
+	BlockBuckets(ctx, s3Client, bucketsToBlock, execute)
 }
