+++ go.work
@@ -1,7 +1,7 @@
 go 1.22.1
 
-use ./bucket-blocker
-
 use ./ingress-inquisition
 
 use ./common
+
+use ./fsbp-fix
